requirements.txt
annotated-types==0.6.0
anyio==4.2.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
beautifulsoup4==4.12.2
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
fastapi==0.108.0
h11==0.14.0
html5lib==1.1
idna==3.6
passlib==1.7.4
pycparser==2.21
pydantic==1.10.12
pydantic_core==2.14.6
PyMySQL==1.1.0
python-dotenv==1.0.0
pytz==2023.3.post1
ratelimit==2.2.1
requests==2.31.0
six==1.16.0
sniffio==1.3.0
soupsieve==2.5
SQLAlchemy==2.0.24
starlette==0.32.0.post1
typing_extensions==4.9.0
urllib3==2.1.0
uvicorn==0.25.0
webencodings==0.5.1
